eth-event-viewer: add initial gall and landscape app
