+++ .urbitrc
@@ -1,5 +1,5 @@
 module.exports = {
   URBIT_PIERS: [
-    "%URBITPIER%",
+    "/home/do7ze5/urbit/binbes-rantem/home",
   ]
 };
