--- full/src/index.css
@@ -1,4 +1,0 @@
-@import 'css/indigo-static.css';
-@import 'css/fonts.css';
-@import 'css/custom.css';
-
