--- full/src/js/components/lib/icons/icon-home.js
@@ -1,15 +1,0 @@
-import React, { Component } from "react";
-
-export class IconHome extends Component {
-  render() {
-    let classes = !!this.props.classes ? this.props.classes : "";
-    return (
-      <img
-        className={"invert-d " + classes}
-        src="/~%APPNAME%/img/Home.png"
-        width={16}
-        height={16}
-      />
-    );
-  }
-}
