--- full/src/js/components/lib/icons/icon-spinner.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-export class IconSpinner extends Component {
-  render() {
-    return (
-      <div className="spinner-pending"></div> 
-    );
-  }
-}
