--- full/src/js/subscription.js
@@ -1,34 +1,0 @@
-import { api } from '/api';
-import { store } from '/store';
-
-import urbitOb from 'urbit-ob';
-
-
-export class Subscription {
-  start() {
-    if (api.authTokens) {
-      // this.initialize%APPNAME%();
-    } else {
-      console.error("~~~ ERROR: Must set api.authTokens before operation ~~~");
-    }
-  }
-
-  // initialize%APPNAME%() {
-  //   api.bind('/primary', 'PUT', api.authTokens.ship, '%APPNAME%',
-  //     this.handleEvent.bind(this),
-  //     this.handleError.bind(this));
-  // }
-
-  handleEvent(diff) {
-    store.handleEvent(diff);
-  }
-
-  handleError(err) {
-    console.error(err);
-    api.bind('/primary', 'PUT', api.authTokens.ship, '%APPNAME%',
-      this.handleEvent.bind(this),
-      this.handleError.bind(this));
-  }
-}
-
-export let subscription = new Subscription();
