--- full/urbit/app/smol/index.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>%APPNAME%</title>
-    <meta charset="utf-8" />
-    <meta name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"/>
-      <link rel="stylesheet" href="/~%APPNAME%/css/index.css" />
-  </head>
-  <body>
-    <div id="root" />
-    <script src="/~/channel/channel.js"></script>
-    <script src="/~modulo/session.js"></script>
-    <script src="/~%APPNAME%/js/index.js"></script>
-  </body>
-</html>
