+++ src/js/reducers/contracts.js
@@ -1,0 +1,13 @@
+import _ from 'lodash';
+
+
+export class ContractsReducer {
+    reduce(json, state) {
+        let data = json;
+        console.log('received data', data);
+        if (data) {
+            state.contracts = data.contracts || state.contracts;
+            state.contract = data.contract || state.contracts;
+        }
+    }
+}
